+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/MoshiChatApi.kt
@@ -217,7 +217,8 @@ constructor(
             connectionId = connectionId,
             request = AddDeviceRequest(
                 device.token,
-                device.pushProvider.key
+                device.pushProvider.key,
+                device.providerName,
             ),
         ).toUnitCall()
     }
@@ -670,6 +671,7 @@ constructor(
             watcherCount = response.watcher_count
             read = response.read.map(DownstreamChannelUserRead::toDomain)
             members = response.members.map(DownstreamMemberDto::toDomain)
+            membership = response.membership?.toDomain()
             messages = response.messages.map { it.toDomain().enrichWithCid(cid) }
             watchers = response.watchers.map(DownstreamUserDto::toDomain)
             hidden = response.hidden