+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/ChannelDtos.kt
@@ -75,6 +75,7 @@ internal data class DownstreamChannelDto(
     val cooldown: Int = 0,
     val pinned_messages: List<DownstreamMessageDto> = emptyList(),
     val own_capabilities: List<String> = emptyList(),
+    val membership: DownstreamMemberDto?,
 
     val extraData: Map<String, Any>,
 ) : ExtraDataDto