+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/di/BaseChatModule.kt
@@ -56,7 +56,7 @@ import io.getstream.chat.android.client.parser.ChatParser
 import io.getstream.chat.android.client.parser2.MoshiChatParser
 import io.getstream.chat.android.client.plugins.requests.ApiRequestsAnalyser
 import io.getstream.chat.android.client.socket.ChatSocket
-import io.getstream.chat.android.client.socket.ChatSocketImpl
+import io.getstream.chat.android.client.socket.SocketFactory
 import io.getstream.chat.android.client.token.TokenManager
 import io.getstream.chat.android.client.token.TokenManagerImpl
 import io.getstream.chat.android.client.uploader.FileUploader
@@ -212,12 +212,13 @@ internal open class BaseChatModule(
         chatConfig: ChatClientConfig,
         parser: ChatParser,
     ): ChatSocket {
-        return ChatSocketImpl(
+        return ChatSocket(
             chatConfig.apiKey,
             chatConfig.wssUrl,
             tokenManager,
-            parser,
+            SocketFactory(parser, tokenManager),
             NetworkStateProvider(appContext.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager),
+            parser,
             networkScope,
         )
     }