+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/decorator/internal/FootnoteDecorator.kt
@@ -227,7 +227,12 @@ internal class FootnoteDecorator(
     }
 
     private fun setupSimpleFootnote(footnoteView: FootnoteView, data: MessageListItem.MessageItem) {
-        footnoteView.showSimpleFootnote()
+        if (data.showMessageFooter) {
+            footnoteView.showSimpleFootnote()
+        } else {
+            footnoteView.hideSimpleFootnote()
+            return
+        }
         setupMessageFooterLabel(footnoteView.footerTextLabel, data, listViewStyle.itemStyle)
         setupMessageFooterTime(footnoteView, data)
         setupDeliveryStateIndicator(footnoteView, data)
@@ -302,7 +307,7 @@ internal class FootnoteDecorator(
         val updatedAt = data.message.getUpdatedAtOrNull()
 
         when {
-            data.isNotBottomPosition() || createdAt == null -> footnoteView.hideTimeLabel()
+            createdAt == null || !data.showMessageFooter -> footnoteView.hideTimeLabel()
             data.message.isGiphyNotEphemeral() && updatedAt != null -> footnoteView.showTime(
                 dateFormatter.formatTime(
                     updatedAt